Move Triggers.h to Common/Constants
