+++ Detectors/Base/CMakeLists.txt
@@ -32,7 +32,6 @@ o2_target_root_dictionary(DetectorsBase
                                   include/DetectorsBase/GeometryManager.h
                                   include/DetectorsBase/MaterialManager.h
                                   include/DetectorsBase/Propagator.h
-                                  include/DetectorsBase/Triggers.h
                                   include/DetectorsBase/Ray.h
                                   include/DetectorsBase/MatCell.h
                                   include/DetectorsBase/MatLayerCyl.h
